feat(frontend): change landing-page to map
This is the main part of out application,
so it should be the first page that is shown to the user.
