+++ frontend/src/routes/index.tsx
@@ -1,54 +1,63 @@
-import { component$ } from "@builder.io/qwik";
+import { $, type Signal, component$, useOn, useSignal } from "@builder.io/qwik";
 import type { DocumentHead } from "@builder.io/qwik-city";
+import { HiFunnelMini, HiXMarkOutline } from "@qwikest/icons/heroicons";
+import { hasVisited, setHasVisited } from "~/components/hasVisited";
+import { IconLinkButton } from "~/components/link_button";
+import { Map } from "~/views/map";
+import { MapUI } from "~/components/map-ui/map-ui";
+import { LoginAvatar } from "~/components/authenticate/authAvatarNav";
 
 export default component$(() => {
+  const organizationLocations: Signal<GeoJSON.GeoJSON> = useSignal({
+    type: "FeatureCollection",
+    features: [],
+  });
+  const projectLocations: Signal<GeoJSON.GeoJSON> = useSignal({
+    type: "FeatureCollection",
+    features: [],
+  });
+
+  // Check if user has visited (by key saved in local storage)
+  // Every visit counts as visit
+  // Don't show by default to prevent short flicker when user has already visited.
+  const showWelcomeMessage = useSignal(false);
+  useOn(
+    "qvisible",
+    $(() => {
+      showWelcomeMessage.value = !hasVisited();
+      setHasVisited(true);
+    }),
+  );
+
   return (
-    <div class="bg-base flex h-screen w-screen flex-col items-center justify-center">
-      <h1 class="mb-8 text-center text-5xl font-bold text-primary">
-        Willkommen zu Sprout!
-      </h1>
-      <h2 class="mx-48 mb-8 text-center text-3xl font-bold text-primary">
-        Möchtest du direkt zur Karte oder zuerst deine Präferenzen angeben,
-        damit wir dich bei deiner Suche unterstützen können?
-      </h2>
-      <div class="flex w-3/4 space-x-16">
-        <a
-          href="/questions"
-          class="transform-all card grow bg-primary transition-all hover:scale-105 active:scale-[1.02]"
-        >
-          <h3 class="my-2 text-center text-xl font-bold text-primary-content">
-            Präferenzen angeben
-          </h3>
-          <figure>
-            <img
-              src="https://picsum.photos/300"
-              alt=""
-              class="h-full w-full"
-              height="4096"
-              width="4096"
-            />
-          </figure>
-        </a>
-        <div class="divider divider-primary divider-horizontal">ODER</div>
-        <a
-          href="/map"
-          class="transform-all card grow bg-primary transition-all hover:scale-105 active:scale-[1.02]"
-        >
-          <h3 class="my-2 text-center text-xl font-bold text-primary-content">
-            Direkt zur Karte
-          </h3>
-          <figure>
-            <img
-              src="https://picsum.photos/300"
-              alt=""
-              class="h-full w-full"
-              height="4096"
-              width="4096"
-            />
-          </figure>
-        </a>
+    <>
+      <Map
+        organizationLocations={organizationLocations}
+        projectLocations={projectLocations}
+      />
+      <MapUI
+        organizationLocations={organizationLocations}
+        projectLocations={projectLocations}
+      />
+      {showWelcomeMessage.value && (
+        <div class="fixed bottom-0 left-0 m-8 rounded bg-base-100 p-8 text-justify">
+          <IconLinkButton
+            target={$(() => (showWelcomeMessage.value = false))}
+            class="absolute right-2 top-2"
+            Icon={HiXMarkOutline}
+          />
+          <h1 class="font-bold">Hey!</h1>
+          <p>Sieht so aus, als hättest du noch keine Schwerpunkte gesetzt.</p>
+          <p>
+            Schau doch mal im <HiFunnelMini class="mx-1 inline" /> Filter-Menü
+            vorbei, um die Ergebnisse etwas einzuschränken.
+          </p>
+        </div>
+      )}
+      <div class="fixed right-6 top-6 rounded-full">
+        <LoginAvatar />
       </div>
-    </div>
+    </>
   );
 });
